+++ src/gh_profiler/cli.py
@@ -1,18 +1,24 @@
 """Main CLI entry point for gh-profiler."""
 
 import sys
+from urllib.parse import urlparse
 
 import click
 
 from . import gh_profiler
 from .utils import cli_utils
 from .utils.profile_data import profile_data as pdata
+from .utils.repo_data import repo_data
+from .utils.cli_config import cli_config
 
 
 @click.command()
 @click.argument("target", required=False)
 @click.version_option(package_name="gh-profiler")
 @click.option("--concise", is_flag=True, help="Show concise output; one flag per category.")
+@click.option("-n", "--num-targets", default=10, help="Preview: How many PRs to review?")
+# @click.option("--issues", is_flag=True, help="Profile contributors of issues rather than PRs.")
+@click.option("--back", is_flag=True, help="Look back over recently merged and closed PRs.")
 @click.option(
     "--generate-workflow",
     is_flag=True,
@@ -21,7 +27,7 @@ from .utils.profile_data import profile_data as pdata
 @click.option("-v", "--verbose", is_flag=True, help="Verbose output, including explanations for decisions about flags.")
 @click.option("--redact", is_flag=True, help="Redact identifying information.")
 @click.option("--benchmark-fetch", is_flag=True, help="Benchmark the code that fetches external data.")
-def main(target, concise, generate_workflow, verbose, redact, benchmark_fetch):
+def main(target, concise, num_targets, back, generate_workflow, verbose, redact, benchmark_fetch):
     """Examine a GitHub user's profile, to help quickly decide how much to invest in their contributions.
 
     You can target a GitHub username, or a PR/issue number from the repository you're working in.
@@ -36,6 +42,11 @@ def main(target, concise, generate_workflow, verbose, redact, benchmark_fetch):
     $ gh-profiler ehmatthes
     $ python -m gh_profiler ehmatthes
       ...
+
+    \b
+    Bulk concise profiling of most recently opened PRs:
+    $ gh-profiler <repo-url>    # 10 most recent PRs
+    $ gh-profiler <repo-url> -n 3
     """
     _validate_command(target, concise, generate_workflow, verbose, redact, benchmark_fetch)
 
@@ -50,6 +61,13 @@ def main(target, concise, generate_workflow, verbose, redact, benchmark_fetch):
     if generate_workflow:
         gh_profiler.main()
         sys.exit()
+
+    # Check if we're processing a repo URL.
+    if "github.com" in target:
+        cli_config.url = target
+        _parse_repo_options(num_targets, back)
+        _parse_repo_info(target)
+        gh_profiler.profile_url()
 
     # If the main argument is an integer, process the PR/issue number.
     # Otherwise, assume it's the username.
@@ -77,3 +95,19 @@ def _validate_command(target, concise, generate_workflow, verbose, redact, bench
     if target and generate_workflow:
         msg = "Please either include a target or --generate-workflow, but not both."
         sys.exit(msg)
+
+def _parse_repo_options(num_targets, back):
+    """Get options relevant to targeting a repo URL."""
+    cli_config.num_targets = num_targets
+    # cli_config.issues = issues
+    cli_config.back = back
+
+def _parse_repo_info(target):
+    """Parse info about the repo from the target.
+
+    Looking for owner and repo name.
+    """
+    parsed_url = urlparse(target)
+    url_parts = parsed_url.path.split("/")
+    repo_data.owner = url_parts[1]
+    repo_data.repo_name = url_parts[2]