+++ src/gh_profiler/utils/repo_summary.py
@@ -1,0 +1,9 @@
+"""Utils for summarizing bulk PRs."""
+
+
+
+
+def show_summary(target_prs):
+    """Show summary for all processed PRs."""
+    for pr in target_prs:
+        print(pr.summary)