+++ tests/e2e_tests/test_profiler.py
@@ -5,8 +5,8 @@ Makes an actual gh-profiler call. Currently calls against my own user account
 to use.
 """
 
-import sys
 import subprocess
+import re
 
 import pytest
 
@@ -129,3 +129,19 @@ def test_redact():
 
     # Should currently see 7 redacted fields in my output.
     assert output.count("<redacted>") == 7
+
+def test_bulk_open_prs():
+    """Test a run against a repo URL for bulk processing open PRs."""
+    # Django is likely to have many open PRs.
+    url = "https://github.com/django/django"
+    cmd = f"uv run gh-profiler {url} -n 3"
+    output = run_with_timeout(cmd)
+
+    assert output.count("https://github.com/django/django/pull/") == 3
+    assert output.count("GitHub user: ") == 3
+
+    # Check that there are 3 PR numbers and titles in output.
+    re_pr_title = r"(PR \d+: ).*"
+    matches = re.findall(re_pr_title, output)
+    assert len(matches) == 3
+